Collapse .h and .cc files in SimG4CMS/EcalTestBeam
