+++ SimG4CMS/EcalTestBeam/plugins/FakeTBEventHeaderProducer.cc
@@ -1,14 +1,40 @@
 /*
  * \file FakeTBEventHeaderProducer.cc
  *
+ * Mimic the event header information
+ * for the test beam simulation
  *
  */
 
 #include "DataFormats/EcalDetId/interface/EBDetId.h"
-#include "SimG4CMS/EcalTestBeam/interface/FakeTBEventHeaderProducer.h"
+#include "DataFormats/Common/interface/Handle.h"
 
-using namespace cms;
-using namespace std;
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/Framework/interface/stream/EDProducer.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+
+#include "Geometry/EcalTestBeam/interface/EcalTBHodoscopeGeometry.h"
+#include "SimDataFormats/EcalTestBeam/interface/PEcalTBInfo.h"
+#include "TBDataFormats/EcalTBObjects/interface/EcalTBEventHeader.h"
+
+class FakeTBEventHeaderProducer : public edm::stream::EDProducer<> {
+public:
+  /// Constructor
+  explicit FakeTBEventHeaderProducer(const edm::ParameterSet &ps);
+
+  /// Destructor
+  ~FakeTBEventHeaderProducer() override;
+
+  /// Produce digis out of raw data
+  void produce(edm::Event &event, const edm::EventSetup &eventSetup) override;
+
+private:
+  edm::EDGetTokenT<PEcalTBInfo> ecalTBInfo_;
+};
 
 FakeTBEventHeaderProducer::FakeTBEventHeaderProducer(const edm::ParameterSet &ps) {
   ecalTBInfo_ = consumes<PEcalTBInfo>(edm::InputTag("EcalTBInfoLabel", "SimEcalTBG4Object"));
@@ -18,7 +44,7 @@ FakeTBEventHeaderProducer::FakeTBEventHeaderProducer(const edm::ParameterSet &ps
 FakeTBEventHeaderProducer::~FakeTBEventHeaderProducer() {}
 
 void FakeTBEventHeaderProducer::produce(edm::Event &event, const edm::EventSetup &eventSetup) {
-  unique_ptr<EcalTBEventHeader> product(new EcalTBEventHeader());
+  std::unique_ptr<EcalTBEventHeader> product(new EcalTBEventHeader());
 
   // get the vertex information from the event
 
@@ -43,8 +69,11 @@ void FakeTBEventHeaderProducer::produce(edm::Event &event, const edm::EventSetup
   product->setTriggerMask(0x1);
   product->setCrystalInBeam(EBDetId(1, theEcalTBInfo->nCrystal(), EBDetId::SMCRYSTALMODE));
 
-  //   LogDebug("FakeTBHeader") << (*product);
-  //   LogDebug("FakeTBHeader") << (*product).eventType();
-  //   LogDebug("FakeTBHeader") << (*product).crystalInBeam();
+  LogDebug("FakeTBHeader") << (*product);
+  LogDebug("FakeTBHeader") << (*product).eventType();
+  LogDebug("FakeTBHeader") << (*product).crystalInBeam();
+
   event.put(std::move(product));
 }
+
+DEFINE_FWK_MODULE(FakeTBEventHeaderProducer);
