+++ SimG4CMS/EcalTestBeam/plugins/TBHodoActiveVolumeRawInfoProducer.cc
@@ -4,9 +4,42 @@
  *
  */
 
-#include "SimG4CMS/EcalTestBeam/interface/TBHodoActiveVolumeRawInfoProducer.h"
+#include "DataFormats/Common/interface/Handle.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/Framework/interface/stream/EDProducer.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+
+#include "Geometry/EcalTestBeam/interface/EcalTBHodoscopeGeometry.h"
+#include "SimDataFormats/CaloHit/interface/PCaloHit.h"
+#include "SimDataFormats/CaloHit/interface/PCaloHitContainer.h"
+#include "SimDataFormats/EcalTestBeam/interface/HodoscopeDetId.h"
+#include "SimDataFormats/EcalTestBeam/interface/PEcalTBInfo.h"
+#include "TBDataFormats/EcalTBObjects/interface/EcalTBHodoscopePlaneRawHits.h"
+#include "TBDataFormats/EcalTBObjects/interface/EcalTBHodoscopeRawInfo.h"
 
 #include <iostream>
+
+class TBHodoActiveVolumeRawInfoProducer : public edm::stream::EDProducer<> {
+public:
+  /// Constructor
+  explicit TBHodoActiveVolumeRawInfoProducer(const edm::ParameterSet &ps);
+
+  /// Destructor
+  ~TBHodoActiveVolumeRawInfoProducer() override;
+
+  /// Produce digis out of raw data
+  void produce(edm::Event &event, const edm::EventSetup &eventSetup) override;
+
+private:
+  double myThreshold;
+
+  edm::EDGetTokenT<edm::PCaloHitContainer> m_EcalToken;
+  EcalTBHodoscopeGeometry *theTBHodoGeom_;
+};
 
 using namespace cms;
 using namespace std;
@@ -83,3 +116,5 @@ void TBHodoActiveVolumeRawInfoProducer::produce(edm::Event &event, const edm::Ev
 
   event.put(std::move(product));
 }
+
+DEFINE_FWK_MODULE(TBHodoActiveVolumeRawInfoProducer);
